Fix offset issue in FileSector Segment.
We weren't accounting for the current segment range assigned to a sector
when segmenting the sector. For this reason, if we had a VFM:

   [0, 300)   Annie.dds   [0, 300)
   [300, 600) Nidalee.dds [0, 300)

And we chopped Nidalee.dds to keep [450, 600), we'd get
   [0, 300)   Annie.dds   [0, 300)
   [450, 600) Nidalee.dds [450, 750)

or something of the like, which is obviously bad.

Also updated the old tests to properly use NMockito.